+++ slips_files/core/database/redis_db/profile_handler.py
@@ -1319,7 +1319,6 @@ class ProfileHandler:
         """
         Used to mark this profile as dhcp server
         """
-
         self.set_profileid_field(profileid, self.constants.GATEWAY, "true")
 
     def set_ipv6_of_profile(self, profileid, ip: list):